Fixed problem and made changes and additions - 3/28/16 12:15PM
I fixed the problem, which was the browser requesting the wrong script file.

I changed the Sprite object constructor's newLoc method. I did a minor change with the
friction so it accepts decimals (0 to 1) and I increased the speed. I put an enemy array
to put all the enemies in and I just added it's image. I drew it too using a for loop.