Adding methods for enemy - 3/28/16 2:30PM
I attempted to add a homing method in the Sprite object constructor, mainly for the enemy. I will
now manually set the speed because it is unfair if the enemy moves as fast as the player,
making no chance of escape available. It didn't work on the first try because in the code,
I had put a comma instead of a period. Then I also messed up the code, which caused the
enemy to go AWAY from the player. It's because of some negative and positive values, so
I just reversed them. And then it worked. I played around with the speed a little. It's
best for the enemy to be half the speed of the player.