Collision method added in Sprite constructor - 3/29/16 9:00PM
I added a colision method in the Sprite constructor called "collidedWith", which takes a
paremeter, a target. It basically checks if it is inside another Sprite, making a collision,
and returning true. Otherwise, false. I played around with it. Also, if the enemy's speed
is equal or greater to the player's speed, the player will never outrun him.