new game.level Array
I had just created a property "level" object in the game object. It contains a start method (puts the player
in it's position, add an enemy, set it's position, add a dog which the player will save, set it's position, etc)
and an update function, which is similar to the updateGame() function, which I removed. I made two simple levels.